+++ CHANGELOG.md
@@ -23,7 +23,8 @@ Changes
 * [REST]: Added a REST API endpoint to read assemblies for samples.
 * [Developer]: `bower` completely removed.
 * [UI/Developer]: Updated analysis results page with Ant Design and React.
-* [UI/Developer]: Updated to `Prettier` 2.0.2.
+* [UI]: Assemblies can now be uploaded through the sample files page.
+* [UI/Developer]: Updated to `Prettier` 2.0.2
 * [UI/Developer]: Updated to `eslint-plugin-react-hooks` to 3.0.0.
 * [UI/Developer]: Removed `angular-ui-bootstrap` as a global dependency (still required on several pages).
 * [UI/Developer]: Added excel file viewer to analysis outputs view.